Final touches

+++ srcs/ft_convert_d_i.c
@@ -6,7 +6,7 @@
 /*   By: cado-car <cado-car@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/08/12 21:22:20 by cado-car          #+#    #+#             */
-/*   Updated: 2021/08/20 16:36:52 by cado-car         ###   ########lyon.fr   */
+/*   Updated: 2021/08/20 16:43:25 by cado-car         ###   ########lyon.fr   */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -28,8 +28,8 @@ static void	ft_convert_d_i_width(t_holder *h, int sign);
 
 void	ft_convert_d_i(t_format *fmt, t_holder *h)
 {
-	int	sign;
-	int	arg;
+	int			sign;
+	long int	arg;
 
 	sign = 1;
 	arg = (int)va_arg(fmt->ap, int);
